refactor: move Annotation to rdf module
